// frontend/src/pages/summary/types/monster.ts
export interface Monster {
  count: number;
  name: string;
  bestiaryLevel: string;
  creatureClass: string;
  deathDmg: number;
  drownDmg: number;
  earthDmg: number;
  energyDmg: number;
  experience: number;
  fireDmg: number;
  hitPoints: number;
  holyDmg: number;
  hpDrainDmg: number;
  iceDmg: number;
  isBoss: boolean;
  physicalDmg: number;
  updatedAt: string;
}